Themes: Add `package.json` switch
Add a `package.json` that sets different entry points on the server and in
the browser (not everything required by `my-sites/themes/index.js` is
SSR-ready yet, so we need a pruned version for the server).
